requirements.txt
fastapi
uvicorn
pydantic
edge-tts
moviepy
pillow
python-multipart